siamrpnretina

+++ mmdet/core/bbox/samplers/base_sampler.py
@@ -71,7 +71,7 @@ class BaseSampler(metaclass=ABCMeta):
             if gt_labels is None:
                 raise ValueError(
                     'gt_labels must be given when add_gt_as_proposals is True')
-            bboxes = torch.cat([gt_bboxes, bboxes], dim=0)
+            bboxes = torch.cat([gt_bboxes, bboxes.double()], dim=0)
             assign_result.add_gt_(gt_labels)
             gt_ones = bboxes.new_ones(gt_bboxes.shape[0], dtype=torch.uint8)
             gt_flags = torch.cat([gt_ones, gt_flags])
